requirements.txt
fastapi
langchain
langchain-openai
tavily-python
uvicorn
python-dotenv
pydantic

